+++ app/src/main/java/nextstep/payments/MainActivity.kt
@@ -3,6 +3,7 @@ package nextstep.payments
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import nextstep.payments.new_card.NewCardScreen
 import nextstep.payments.ui.theme.PaymentsTheme
 
 class MainActivity : ComponentActivity() {
